more logging and checking

+++ create_pending_verses.js
@@ -11,11 +11,13 @@ const urls = fs.readFileSync(logFile).toString().trim().split("\n").forEach(url 
   const verse = urlComps.pop().replace("-","/").replace(".html",".txt");
   const book = urlComps.pop();
   const verseFile = `./data/${book}/${verse}`;
-  axios.get(url).then(response => {
-    const dom = new JSDOM(response.data);
-    const title = dom.window.document.querySelector(".extra_title").textContent;
-    const body = dom.window.document.querySelector(".subtitle").textContent;
-    fs.writeFileSync(verseFile, title + "\n" + body);
-    console.log(`Done creating verse ${verseFile}`);
-  });
+  if (! fs.existsSync(verseFile) ) {
+    axios.get(url).then(response => {
+      const dom = new JSDOM(response.data);
+      const title = dom.window.document.querySelector(".extra_title").textContent;
+      const body = dom.window.document.querySelector(".subtitle").textContent;
+      fs.writeFileSync(verseFile, title + "\n" + body);
+      console.log(`Done creating verse ${verseFile}`);
+    });
+  }
 });
